Move stream to separate mods
