+++ tonic-tls/src/rustls/server.rs
@@ -1,9 +1,32 @@
 use std::sync::Arc;
 
-use crate::rustls::RustlsAcceptor;
-
 use super::TlsStream;
 use futures::Stream;
+use tokio::io::{AsyncRead, AsyncWrite};
+
+/// Internal implementation of acceptor wrapper.
+#[derive(Clone)]
+struct RustlsAcceptor(tokio_rustls::TlsAcceptor);
+
+impl RustlsAcceptor {
+    fn new(inner: tokio_rustls::TlsAcceptor) -> Self {
+        Self(inner)
+    }
+}
+
+impl<S> crate::server::TlsAcceptor<S> for RustlsAcceptor
+where
+    S: AsyncRead + AsyncWrite + Send + Sync + Unpin + 'static,
+{
+    type TlsStream = TlsStream<S>;
+    async fn accept(&self, stream: S) -> Result<TlsStream<S>, crate::Error> {
+        self.0
+            .accept(stream)
+            .await
+            .map(|s| TlsStream::new(s))
+            .map_err(crate::Error::from)
+    }
+}
 
 /// The same as the [incoming] returned stream,
 /// but wrapped in a struct.
@@ -42,7 +65,7 @@ impl TlsIncoming {
     ) -> Self {
         let acceptor = RustlsAcceptor::new(tokio_rustls::TlsAcceptor::from(server_config));
         Self {
-            inner: crate::incoming_inner(tcp_incoming, acceptor),
+            inner: crate::server::incoming_inner(tcp_incoming, acceptor),
         }
     }
 }
